Resolution loop and gaussian weights for the distance.

+++ inc/common.h
@@ -12,7 +12,7 @@ using namespace std;
 int diff(const Vec3b& pa, const Vec3b& pb);
 
 // Difference of two neighborhoods
-int diff_NH(const Point& pa, const Point& pb, const Mat& imagea, const Mat& imageb, const int& w);
+double diff_NH(const Point& pa, const Point& pb, const Mat& imagea, const Mat& imageb, const int& w);
 
 // Compute the neighborhood of a pixel in a given image (neigborhood of size (w+1)*(w+1)
 vector<Vec3b>  neighborhood(const Point& pixel, const int& w, const Mat& image);
